Array-class {S4Arrays}	R Documentation

Array objects

Description

Array is a virtual class with no slots intended to be extended by concrete subclasses with an array-like semantic.

Details

Some examples of Array derivatives:

SparseArray objects implemented in the SparseArray package.

DelayedArray objects implemented in the DelayedArray package.

ArrayGrid and ArrayViewport objects implemented in this package (the S4Arrays package).

See Also

array and matrix objects in base R.

Examples

showClass("Array")  # virtual class with no slots

[Package S4Arrays version 1.10.1 Index]