Array-subassignment {S4Arrays}	R Documentation

Low-level internal generics involved in subassignment of Array derivatives

Description

The S4Arrays package defines a small set of low-level generic functions that are involved in subassignment operations on Array derivatives. They are not intended to be used directly by the end user.

See Also

The Array class.

[Package S4Arrays version 1.10.1 Index]